Fix ImportH5Data class templating for macos & linux
